requirements.txt
python-telegram-bot==21.8
PyYAML>=6.0
pydantic>=2.7,<3.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
